Fixup configure.sh:
This fixes the configure.sh completion message so that the printed "alias"
commands that are intended to be copied into the .bashrc or .cshrc files
actually would work.  It also adds the alias definition to
captain.profile.in, so that even if the alias isn't defined in the
login files, it will be defined after the captain.profile has
been sourced.