Just use "master" as the default installation version.
I was following "big experiment" practice and installing a fixed version by
default, but that doesn't make sense when there are only a handful of
people working on the software.  It's good in principle, but we don't have
enough people to keep up with freezing a stable version.  We will need to
do it once we start with "real analysis", but not while we are still
developing.